+++ plugin-locationcomponent/src/test/java/com/mapbox/maps/plugin/locationcomponent/LayerSourceProviderTest.kt
@@ -70,6 +70,7 @@ class LayerSourceProviderTest {
         "model-cast-shadows" to true,
         "model-scale-transition" to hashMapOf("duration" to 0, "delay" to 0),
         "model-rotation-transition" to hashMapOf("duration" to 0, "delay" to 0),
+        "model-elevation-reference" to "ground",
         "model-emissive-strength" to 1.0,
         "model-color" to listOf("rgba", 255, 255, 255, 1.0),
         "model-color-mix-intensity" to 0.0,